Consolidate VisualGen LPIPS eval script
